pyproject.toml
[project]
name = "vax"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Calvin Laurenson", email = "hello@laurenson.dev" }
]
dependencies = [
    "jax>=0.4.33",
    "flax>=0.9.0",
    "rich>=13.8.1",
    "click>=8.1.7",
    "rich-click>=1.8.3",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "opencv-python>=4.10.0.84",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
vax = "vax.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vax"]

[tool.ruff]
line-length = 120
